requirements.txt
pandas~=1.0.5
numpy~=1.18.5
matplotlib~=3.2.2
scikit-image~=0.16.2
scikit-learn~=0.23.1
tqdm~=4.47.0
torchvision~=0.7.0
scipy~=1.4.1
pillow~=7.2.0
networkx~=2.4
torch~=1.6.0